refactor(typescript): move cript to types/cript
