+++ scripts/typescript/src/types/cript/IAlgorithm.ts
