+++ scripts/typescript/src/types/cript/ICitation.ts
