+++ scripts/typescript/src/types/cript/ICollection..ts
