+++ scripts/typescript/src/types/cript/IComputationProcess.ts
