+++ scripts/typescript/src/types/cript/ICondition.d.ts
