+++ scripts/typescript/src/types/cript/IData.ts
