+++ scripts/typescript/src/types/cript/IEquipment.ts
