+++ scripts/typescript/src/types/cript/IExperiment.ts
