+++ scripts/typescript/src/types/cript/IFile.ts
