+++ scripts/typescript/src/types/cript/IIdentifier.ts
