+++ scripts/typescript/src/types/cript/IIngredient.ts
