+++ scripts/typescript/src/types/cript/IInventory.ts
