+++ scripts/typescript/src/types/cript/IMaterial.ts
