+++ scripts/typescript/src/types/cript/IParameter.ts
