+++ scripts/typescript/src/types/cript/IProcess.ts
