+++ scripts/typescript/src/types/cript/IProject.ts
