+++ scripts/typescript/src/types/cript/IProperty.ts
