+++ scripts/typescript/src/types/cript/IQuantity.ts
