+++ scripts/typescript/src/types/cript/IReference.ts
