+++ scripts/typescript/src/types/cript/ISoftware.ts
