+++ scripts/typescript/src/types/cript/ISoftwareConfiguration.ts
