+++ scripts/typescript/src/types/cript/IUser.ts
