+++ scripts/typescript/src/types/cript/IVocab.ts
