+++ scripts/typescript/src/types/cript/index.ts
